requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
boto3==1.35.95
botocore==1.35.95
celery==5.5.3
channels==4.2.0
channels-redis==4.2.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
daphne==4.1.2
Django==5.2.8
jmespath==1.0.1
kombu==5.5.4
numpy==2.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
smart-open[s3]==7.1.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
